PP-5322: restructure github strategy, permissions
* move github specific strategy and permissions into their own folder
* cleanup comments and respond to review
